Adds use of sliceExists in release.py script

+++ .ci/release.py
@@ -5,4 +5,8 @@ from geni import util
 
 
 ctx = util.loadContext(key_passphrase=os.environ['GENI_KEY_PASSPHRASE'])
-util.deleteSliverExists(cloudlab.Clemson, ctx, 'popperized')
+
+if util.sliceExists(ctx, 'popperized'):
+    util.deleteSliverExists(cloudlab.Clemson, ctx, 'popperized')
+else:
+    print("Slice 'popperized' does not exist")
